cleanup: isolate peer delivery helpers
